refactor(runtime): Re-home runtime core nucleus
Move the remaining runtime bridge classes and package-local tests from network.crypta.node.runtime to network.crypta.runtime.core.

Update direct callers and runtime-admin documentation to use the new package without changing build structure or runtime behavior.
