+++ src/main/java/network/crypta/clients/fcp/FcpServerConfigRegistrar.java
@@ -10,7 +10,7 @@ import network.crypta.config.StringCallback;
 import network.crypta.config.SubConfig;
 import network.crypta.io.NetworkInterface;
 import network.crypta.l10n.NodeL10n;
-import network.crypta.node.runtime.SSL;
+import network.crypta.runtime.core.SSL;
 
 /**
  * Registers FCP server configuration options and constructs a configured {@link FCPServer}.
