+++ src/main/java/network/crypta/clients/http/SimpleToadletServer.java
@@ -32,8 +32,8 @@ import network.crypta.keys.FreenetURI;
 import network.crypta.l10n.NodeL10n;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.PrioRunnable;
-import network.crypta.node.runtime.SSL;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.core.SSL;
 import network.crypta.runtime.spi.RandomnessPort;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.HTMLNode;
