+++ src/main/java/network/crypta/clients/http/ToadletContextImpl.java
@@ -28,8 +28,8 @@ import java.util.concurrent.atomic.AtomicReference;
 import network.crypta.clients.http.FProxyFetchInProgress.REFILTER_POLICY;
 import network.crypta.clients.http.bookmark.BookmarkManager;
 import network.crypta.l10n.NodeL10n;
-import network.crypta.node.runtime.SSL;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.core.SSL;
 import network.crypta.support.HTMLEncoder;
 import network.crypta.support.HTMLNode;
 import network.crypta.support.MultiValueTable;
