+++ src/main/java/network/crypta/runtime/core/LegacyConfigPort.java
@@ -1,4 +1,4 @@
-package network.crypta.node.runtime;
+package network.crypta.runtime.core;
 
 import java.util.EnumMap;
 import java.util.LinkedHashMap;
