+++ src/main/java/network/crypta/runtime/core/LegacyConnectivityPort.java
@@ -1,4 +1,4 @@
-package network.crypta.node.runtime;
+package network.crypta.runtime.core;
 
 import java.util.ArrayList;
 import java.util.List;
