+++ src/main/java/network/crypta/runtime/core/LegacyNodeInfoPort.java
@@ -1,4 +1,4 @@
-package network.crypta.node.runtime;
+package network.crypta.runtime.core;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
