+++ src/test/java/network/crypta/node/NodeClientCoreTest.java
@@ -19,9 +19,9 @@ import network.crypta.crypt.MasterSecret;
 import network.crypta.crypt.RandomSource;
 import network.crypta.keys.NodeSSK;
 import network.crypta.node.SecurityLevels.PHYSICAL_THREAT_LEVEL;
-import network.crypta.node.runtime.LegacyRuntimePorts;
 import network.crypta.node.subsystem.NodeServicesSubsystem;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.core.LegacyRuntimePorts;
 import network.crypta.runtime.endpoints.ClientEndpoints;
 import network.crypta.runtime.endpoints.NodeClientPersistence;
 import network.crypta.runtime.spi.RuntimePorts;
