+++ src/test/java/network/crypta/node/NodeStarterTest.java
@@ -12,8 +12,8 @@ import network.crypta.config.SubConfig;
 import network.crypta.crypt.CryptoRandoms;
 import network.crypta.crypt.DummyRandomSource;
 import network.crypta.crypt.RandomSource;
-import network.crypta.node.runtime.SSL;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
+import network.crypta.runtime.core.SSL;
 import network.crypta.support.SimpleFieldSet;
 import network.crypta.support.io.NativeThread;
 import org.junit.jupiter.api.AfterEach;
