+++ src/test/java/network/crypta/runtime/core/LegacyConfigPortTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node.runtime;
+package network.crypta.runtime.core;
 
 import java.util.EnumSet;
 import java.util.Map;
