+++ src/test/java/network/crypta/runtime/core/LegacyConnectivityPortTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node.runtime;
+package network.crypta.runtime.core;
 
 import java.net.InetAddress;
 import network.crypta.config.PersistentConfig;
