+++ src/test/java/network/crypta/runtime/core/LegacyPeerPortTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node.runtime;
+package network.crypta.runtime.core;
 
 import java.util.List;
 import java.util.Map;
