+++ src/test/java/network/crypta/runtime/core/LegacyRuntimePortsTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node.runtime;
+package network.crypta.runtime.core;
 
 import java.io.File;
 import java.util.Random;
